requirements.txt
openai
web3
argparse
pygame
speechrecognition
gtts
google-api-python-client
requests
pillow
cirq
numpy